MBP-006: refactored views to containers, moved component logic to container file
