+++ frontend/blog/src/components/NavbarComponent.vue
@@ -1,4 +1,3 @@
-
 <script setup>
 
 </script>
