--- frontend/blog/src/views/PostDetailView.vue
@@ -1,15 +1,0 @@
-<template>
-    <PostDetailComponent></PostDetailComponent>
-</template>
-
-<script>
-import PostDetailComponent from '@/components/PostDetailComponent'
-
-
-export default {
-  name: 'PostDetailView',
-  components: {
-    PostDetailComponent
-  },
-}
-</script>
